+++ servicenotifier/pom.xml
